+++ scripts/runtest.sh
@@ -22,7 +22,7 @@ then
     VERBOSE=1
 fi
 
-accepted_commands=(static short acceptance sequential-acceptance multiple binary 
+accepted_commands=(static token short acceptance sequential-acceptance multiple binary
     binary-prepare catalog gateway vapp vm network extnetwork multinetwork 
     short-provider lb user acceptance-orguser short-provider-orguser)
 
@@ -194,6 +194,50 @@ function exists_in_path {
     done
 }
 
+function make_token {
+  jq=$(exists_in_path jq)
+  if [ -z "$jq" ]
+  then
+    echo "program jq not found"
+    exit 1
+  fi
+  curl=$(exists_in_path curl)
+  if [ -z "$curl" ]
+  then
+    echo "program curl not found"
+    exit 1
+  fi
+  check_for_config_file
+  echo "# Using credentials from $config_file"
+  user=$(jq -r '.provider.user' $config_file)
+  password=$(jq -r '.provider.password' $config_file)
+  url=$(jq -r '.provider.url' $config_file)
+  org=$(jq -r '.provider.sysOrg' $config_file)
+
+  if [ -z "$user" -o "$user" == "null" ]
+  then
+    echo "missing user from configuration file. Can't retrieve token"
+    exit 1
+  fi
+  if [ -z "$password" -o "$password" == "null" ]
+  then
+    echo "missing password from configuration file. Can't retrieve token"
+    exit 1
+  fi
+  if [ -z "$org" -o "$org" == "null" ]
+  then
+    echo "missing org from configuration file. Can't retrieve token"
+    exit 1
+  fi
+  auth=$(echo -n "$user@$org:$password" |base64)
+
+  echo "# Connecting to $url"
+  curl --silent --head --insecure \
+    --header "Accept: application/*;version=27.0" \
+    --header "Authorization: Basic $auth" \
+    --request POST $url/sessions | grep -i authorization
+}
+
 function check_static {
     static_check=$(exists_in_path staticcheck)
     if [  -z "$staticcheck" -a -n "$TRAVIS" ]
@@ -250,6 +294,9 @@ function check_static {
 case $wanted in
     static)
         check_static
+        ;;
+    token)
+        make_token
         ;;
     test-env-init)
         export VCD_ENV_INIT=1