+++ vcd/datasource_vcd_network_test.go
@@ -58,7 +58,7 @@ func getAvailableNetworks() error {
 	if err != nil {
 		return fmt.Errorf("error getting client configuration: %s", err)
 	}
-	err = vcdClient.Authenticate(testConfig.Provider.User, testConfig.Provider.Password, testConfig.Provider.SysOrg)
+	err = ProviderAuthenticate(vcdClient, testConfig.Provider.User, testConfig.Provider.Password, testConfig.Provider.Token, testConfig.Provider.SysOrg)
 	if err != nil {
 		return fmt.Errorf("authentication error: %s", err)
 	}